Refactor to remove separate telemetry initializer
Centralize activity tracing into the middleware itself to avoid
"hackiness" around propagating the `Activity` outside of the pipeline
into the telemetry initializer.
